requirements.txt
google-generativeai
Pillow
opencv-python
matplotlib
numpy
tk